pyproject.toml
[tool.poetry]
name = "moka-util"
version = "0.1.0"
description = ""
authors = ["Ezra Tampubolon <ezra.tampubolon@googlemail.com>"]
readme = "README.md"
packages = [{include = "moka_util"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
black = "^24.10.0"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
isort = "^5.13.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
